requirements/prod.txt
# Production module requirements go here.
# Be sure that you always specify an exact version such as:
# foo==1.2.3 and you list all dependencies explicitly.

#-r compiled.txt

cef==0.5
commonware==0.4.2
Django==1.7.6
django-filter==0.7
django-paranoia==0.1.8.7
django-statsd-mozilla==0.3.10
djangorestframework==2.3.12
hawkrest==0.0.2
mohawk==0.2.1
mozilla-logger==0.2
raven==4.0.3
six==1.4.1
statsd==2.1.2
websigning==0.1
